+++ API/admin.js
@@ -1,0 +1,16 @@
+'use strict';
+
+const routes = require('express').Router(); // eslint-disable-line new-cap
+
+/*
+routes.get('/admin/list_users', (req, res) => {
+	admin.list_users(function(err, data) {
+		if (err) {
+			console.log(err.text);
+		} else {
+			res.status(200).json(data);
+		}
+	});
+});*/
+
+module.exports = routes;
