+++ API/user.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const routes = require('express').Router(); // eslint-disable-line new-cap
+
+/* App-specific module imports */
+
+module.exports = routes;
